final dense to float64

+++ Outcome_Analysis/DeepLearningTools/DenseNetModel/MyDenseNet.py
@@ -228,7 +228,7 @@ def DenseNet(blocks, include_top=False, weights='imagenet', input_shape=(32, 128
     x = layers.Dropout(0.5)(x)
     x = layers.Dense(512, activation='relu', kernel_regularizer=regularizers.l2(0.001))(x)
     x = layers.Dropout(0.5)(x)
-    x = layers.Dense(classes, activation='softmax', name='predictions', dtype='float32')(x)
+    x = layers.Dense(classes, activation='softmax', name='predictions', dtype='float64')(x)
 
     if blocks == [6, 12, 24, 16]:
         model_name = 'densenet121'
